Migrate mega q edge to q edge
